backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==1.0.1
stripe==5.4.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0